requirements.txt
fastapi>=0.111,<1
uvicorn[standard]>=0.30,<1
sqlalchemy>=2.0,<3
pydantic>=2.7,<3
pydantic-settings>=2.3,<3
python-dotenv>=1.0,<2
google-genai>=1.0.0
openai>=1.0.0
pydantic-ai-slim[openai,google]; python_version >= "3.10"
alembic>=1.13,<2
pytest>=8.0,<9
